+++ Runtime/Qr/QrMatrix.cs
@@ -14,8 +14,8 @@ namespace TestingFloor.Internal {
         public bool this[int row, int col] => _modules[row, col];
 
         /// Renders the matrix to a `Texture2D` with `scale`-pixel-wide modules and a `quiet`-module quiet zone.
-        /// Dark = black, light = white.
-        public Texture2D ToTexture(int scale = 1, int quiet = 2) {
+        /// Normal mode: dark = black, light = white. Inverted mode swaps those colors.
+        public Texture2D ToTexture(int scale = 1, int quiet = 2, bool inverted = false) {
             if (scale < 1) scale = 1;
             if (quiet < 0) quiet = 0;
 
@@ -28,8 +28,10 @@ namespace TestingFloor.Internal {
             var colors = new Color32[pixels * pixels];
             var white = new Color32(255, 255, 255, 255);
             var black = new Color32(0, 0, 0, 255);
+            var light = inverted ? black : white;
+            var dark = inverted ? white : black;
 
-            for (var i = 0; i < colors.Length; i++) colors[i] = white;
+            for (var i = 0; i < colors.Length; i++) colors[i] = light;
 
             for (var r = 0; r < Size; r++) {
                 for (var c = 0; c < Size; c++) {
@@ -41,7 +43,7 @@ namespace TestingFloor.Internal {
                         var ty = pixels - 1 - (texRow + dy);
                         var baseIdx = ty * pixels;
                         for (var dx = 0; dx < scale; dx++) {
-                            colors[baseIdx + texCol + dx] = black;
+                            colors[baseIdx + texCol + dx] = dark;
                         }
                     }
                 }